requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
drf-yasg==1.21.7
psycopg2-binary==2.9.7
django-redis==5.2.0
redis==5.0.1
coreapi==2.3.3
Pillow==10.0.1